requirements.txt
openai>=1.0.0
langchain-openai
langchain-core
pywebview

